# apps/engine/src/__generated__/openapi.json
{
  "openapi": "3.1.0",
  "info": {
    "title": "Stripe Sync Engine",
    "version": "1.0.0",
    "description": "Stripe Sync Engine — stateless, one-shot source/destination sync over HTTP.\nAll sync endpoints accept configuration via the `X-Pipeline` header (JSON-encoded PipelineConfig). Optional cursor state can be provided via `X-State`.\n\n## Endpoints\n\n| Method | Path | Summary |\n|--------|------|---------|\n| GET | /health | Health check |\n| POST | /setup | Set up destination schema |\n| POST | /teardown | Tear down destination schema |\n| GET | /check | Check connector connection |\n| POST | /discover | Discover available streams |\n| POST | /read | Read records from source |\n| POST | /write | Write records to destination |\n| POST | /sync | Run sync pipeline (read → write) |\n| GET | /connectors | List available connectors and their config schemas |"
  },
  "paths": {
...
    "/setup": {
      "post": {
        "operationId": "setup",
        "tags": ["Stateless Sync API"],
        "summary": "Set up destination schema",
        "description": "Creates destination tables and applies migrations. Safe to call multiple times.",
        "parameters": [
          {
            "in": "header",
            "name": "x-pipeline",
            "schema": {
              "type": "string",
              "contentMediaType": "application/json",
              "contentSchema": {
                "$ref": "#/components/schemas/PipelineConfig"
              }
            },
            "description": "JSON-encoded PipelineConfig: { source: { type, ...config }, destination: { type, ...config }, streams }"
          }
        ],
        "responses": {
          "200": {
            "description": "Setup complete",
            "content": {
              "application/json": {
                "schema": {
                  "type": "object",
                  "propertyNames": {
                    "type": "string"
                  },
                  "additionalProperties": {}
                }
              }
            }
          },
          "400": {
            "description": "Invalid params",
            "content": {
              "application/json": {
                "schema": {
                  "type": "object",
                  "properties": {
                    "error": {}
                  },
                  "required": ["error"],
                  "additionalProperties": false
                }
              }
            }
          }
        }
      }
    },
    "/teardown": {
      "post": {
        "operationId": "teardown",
        "tags": ["Stateless Sync API"],
        "summary": "Tear down destination schema",
        "description": "Drops destination tables. Irreversible.",
        "parameters": [
          {
            "in": "header",
            "name": "x-pipeline",
            "schema": {
              "type": "string",
              "contentMediaType": "application/json",
              "contentSchema": {
                "$ref": "#/components/schemas/PipelineConfig"
              }
            },
            "description": "JSON-encoded PipelineConfig: { source: { type, ...config }, destination: { type, ...config }, streams }"
          }
        ],
        "responses": {
          "204": {
            "description": "Teardown complete"
          },
          "400": {
            "description": "Invalid params",
            "content": {
              "application/json": {
                "schema": {
                  "type": "object",
                  "properties": {
                    "error": {}
                  },
                  "required": ["error"],
                  "additionalProperties": false
                }
              }
            }
          }
        }
      }
    },
    "/check": {
      "get": {
        "operationId": "check",
        "tags": ["Stateless Sync API"],
        "summary": "Check connector connection",
        "description": "Validates the source/destination config and tests connectivity.",
        "parameters": [
          {
            "in": "header",
            "name": "x-pipeline",
            "schema": {
              "type": "string",
              "contentMediaType": "application/json",
              "contentSchema": {
                "$ref": "#/components/schemas/PipelineConfig"
              }
            },
            "description": "JSON-encoded PipelineConfig: { source: { type, ...config }, destination: { type, ...config }, streams }"
          }
        ],
        "responses": {
          "200": {
            "description": "Connection check result",
            "content": {
              "application/json": {
                "schema": {
                  "type": "object",
                  "properties": {
                    "source": {
                      "type": "object",
                      "properties": {
                        "status": {
                          "type": "string",
                          "enum": ["succeeded", "failed"]
                        },
                        "message": {
                          "type": "string"
                        }
                      },
                      "required": ["status"],
                      "additionalProperties": false
                    },
                    "destination": {
                      "type": "object",
                      "properties": {
                        "status": {
                          "type": "string",
                          "enum": ["succeeded", "failed"]
                        },
                        "message": {
                          "type": "string"
                        }
                      },
                      "required": ["status"],
                      "additionalProperties": false
                    }
                  },
                  "required": ["source", "destination"],
                  "additionalProperties": false
                }
              }
            }
          },
          "400": {
            "description": "Invalid params",
            "content": {
              "application/json": {
                "schema": {
                  "type": "object",
                  "properties": {
                    "error": {}
                  },
                  "required": ["error"],
                  "additionalProperties": false
                }
              }
            }
          }
        }
      }
    },
    "/discover": {
      "post": {
        "operationId": "discover",
        "tags": ["Stateless Sync API"],
        "summary": "Discover available streams",
        "description": "Returns the catalog of available streams for the configured source. Each stream includes its name, primary key, and optional JSON schema.",
        "parameters": [
          {
            "in": "header",
            "name": "x-pipeline",
            "schema": {
              "type": "string",
              "contentMediaType": "application/json",
              "contentSchema": {
                "$ref": "#/components/schemas/PipelineConfig"
              }
            },
            "description": "JSON-encoded PipelineConfig: { source: { type, ...config }, destination: { type, ...config }, streams }"
          }
        ],
        "responses": {
          "200": {
            "description": "Available streams",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/CatalogMessageOutput"
                }
              }
            }
          },
          "400": {
            "description": "Invalid params",
            "content": {
              "application/json": {
                "schema": {
                  "type": "object",
                  "properties": {
                    "error": {}
                  },
                  "required": ["error"],
                  "additionalProperties": false
                }
              }
            }
          }
        }
      }
    },
    "/read": {
      "post": {
        "operationId": "read",
        "tags": ["Stateless Sync API"],
        "summary": "Read records from source",
        "description": "Streams NDJSON messages (records, state, catalog). Optional NDJSON body provides live events as input.",
        "parameters": [
          {
            "in": "header",
            "name": "x-pipeline",
            "schema": {
              "type": "string",
              "contentMediaType": "application/json",
              "contentSchema": {
                "$ref": "#/components/schemas/PipelineConfig"
              }
            },
            "description": "JSON-encoded PipelineConfig: { source: { type, ...config }, destination: { type, ...config }, streams }"
          },
          {
            "in": "header",
            "name": "x-state",
            "schema": {
              "type": "string",
              "contentMediaType": "application/json",
              "contentSchema": {
                "type": "object",
                "additionalProperties": true,
                "description": "Per-stream cursor state keyed by stream name"
              }
            },
            "description": "JSON-encoded per-stream cursor state. Engine uses this if present, falls back to StateStore."
          },
          {
            "in": "header",
            "name": "x-state-checkpoint-limit",
            "schema": {
              "description": "When set, stops streaming after N state checkpoint messages. Enables page-at-a-time sync.",
              "example": "1",
              "type": "integer",
              "exclusiveMinimum": 0,
              "maximum": 9007199254740991
            },
            "description": "When set, stops streaming after N state checkpoint messages. Enables page-at-a-time sync."
          }
        ],
        "responses": {
          "200": {
            "description": "NDJSON stream of sync messages",
            "content": {
              "application/x-ndjson": {
                "schema": {
                  "$ref": "#/components/schemas/MessageOutput"
                }
              }
            }
          },
          "400": {
            "description": "Invalid params",
            "content": {
              "application/json": {
                "schema": {
                  "type": "object",
                  "properties": {
                    "error": {}
                  },
                  "required": ["error"],
                  "additionalProperties": false
                }
              }
            }
          }
        }
      }
    },
    "/write": {
      "post": {
        "operationId": "write",
        "tags": ["Stateless Sync API"],
        "summary": "Write records to destination",
        "description": "Reads NDJSON messages from the request body and writes them to the destination. Pipe /read output as input.",
        "parameters": [
          {
            "in": "header",
            "name": "x-pipeline",
            "schema": {
              "type": "string",
              "contentMediaType": "application/json",
              "contentSchema": {
                "$ref": "#/components/schemas/PipelineConfig"
              }
            },
            "description": "JSON-encoded PipelineConfig: { source: { type, ...config }, destination: { type, ...config }, streams }"
          }
        ],
        "requestBody": {
          "required": true,
          "content": {
            "application/x-ndjson": {
              "schema": {
                "$ref": "#/components/schemas/Message"
              }
            }
          }
        },
        "responses": {
          "200": {
            "description": "NDJSON stream of write result messages",
            "content": {
              "application/x-ndjson": {
                "schema": {
                  "$ref": "#/components/schemas/DestinationOutput"
                }
              }
            }
          },
          "400": {
            "description": "Invalid params",
            "content": {
              "application/json": {
                "schema": {
                  "type": "object",
                  "properties": {
                    "error": {}
                  },
                  "required": ["error"],
                  "additionalProperties": false
                }
              }
            }
          }
        }
      }
    },
...
  },
  "components": {
    "schemas": {
...
      "CatalogMessageOutput": {
        "type": "object",
        "properties": {
          "type": {
            "type": "string",
            "const": "catalog"
          },
          "streams": {
            "type": "array",
            "items": {
              "type": "object",
              "properties": {
                "name": {
                  "type": "string"
                },
                "primary_key": {
                  "type": "array",
                  "items": {
                    "type": "array",
                    "items": {
                      "type": "string"
                    }
                  }
                },
                "json_schema": {
                  "type": "object",
                  "propertyNames": {
                    "type": "string"
                  },
                  "additionalProperties": {}
                },
                "metadata": {
                  "type": "object",
                  "propertyNames": {
                    "type": "string"
                  },
                  "additionalProperties": {}
                }
              },
              "required": ["name", "primary_key"],
              "additionalProperties": false
            }
          }
        },
        "required": ["type", "streams"],
        "additionalProperties": false
      },
...
      "PipelineConfig": {
        "type": "object",
        "required": ["source", "destination"],
        "properties": {
          "source": {
            "type": "object",
            "required": ["type"],
            "properties": {
              "type": {
                "type": "string"
              }
            },
            "additionalProperties": true
          },
          "destination": {
            "type": "object",
            "required": ["type"],
            "properties": {
              "type": {
                "type": "string"
              }
            },
            "additionalProperties": true
          },
          "streams": {
            "type": "array",
            "items": {
              "type": "object",
              "required": ["name"],
              "properties": {
                "name": {
                  "type": "string"
                },
                "sync_mode": {
                  "type": "string",
                  "enum": ["incremental", "full_refresh"]
                },
                "fields": {
                  "type": "array",
                  "items": {
                    "type": "string"
                  }
                }
              }
            }
          }
        }
      }
    }
  }
}